Add files via upload
This wad is for the WII and is a channel for HOLES the movie by Louis Sachar. this wad is a modified version of the homer channel credit to the person who made the homer channel. use at your own risk! I am not responsible if your WII or anything else crashes/bricks your WII or device and would be due to the negligence of the person who put this WAD on there WII. This is a beta version of the WAD and some day there might be a game I  make for the channel. And as always, have fun and stay safe!